merge branch integration_2026-06-04_1160742843906 into master

See merge request: !1155

+++ service/fwcenter/api_describe_control_policy_by_rule_id.go
@@ -180,7 +180,7 @@ type DataForDescribeControlPolicyByRuleIdOutput struct {
 
 	EffectStatus *int32 `type:"int32" json:",omitempty"`
 
-	EndTime *int32 `type:"int32" json:",omitempty"`
+	EndTime *int64 `type:"int64" json:",omitempty"`
 
 	HitCnt *int64 `type:"int64" json:",omitempty"`
 
@@ -334,7 +334,7 @@ func (s *DataForDescribeControlPolicyByRuleIdOutput) SetEffectStatus(v int32) *D
 }
 
 // SetEndTime sets the EndTime field's value.
-func (s *DataForDescribeControlPolicyByRuleIdOutput) SetEndTime(v int32) *DataForDescribeControlPolicyByRuleIdOutput {
+func (s *DataForDescribeControlPolicyByRuleIdOutput) SetEndTime(v int64) *DataForDescribeControlPolicyByRuleIdOutput {
 	s.EndTime = &v
 	return s
 }
